+++ src/main/java/net/frozenblock/wilderwild/mixin/projectile/ThrownEnderpearlMixin.java
@@ -23,7 +23,6 @@ import net.frozenblock.wilderwild.config.ItemConfig;
 import net.frozenblock.wilderwild.registry.RegisterSounds;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvents;
-import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.projectile.ThrownEnderpearl;
 import net.minecraft.world.phys.HitResult;
@@ -31,36 +30,45 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 @Mixin(ThrownEnderpearl.class)
 public class ThrownEnderpearlMixin {
 
-	@Inject(method = "onHit", at = @At("HEAD"))
-	public void wilderWild$onHit(HitResult result, CallbackInfo info) {
+	@Inject(
+		method = "onHit",
+		at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;resetFallDistance()V", ordinal = 0),
+		locals = LocalCapture.CAPTURE_FAILHARD
+	)
+	public void wilderWild$onHitWithServerPlayer(HitResult result, CallbackInfo info, Entity entity) {
+		if (ItemConfig.get().projectileLandingSounds.enderPearlLandingSounds && entity instanceof ServerPlayer owner) {
+			ThrownEnderpearl pearl = ThrownEnderpearl.class.cast(this);
+			if (!pearl.isSilent()) {
+				float pitch = 0.9F + (pearl.level().random.nextFloat() * 0.2F);
+				pearl.level().playSound(owner, pearl.getX(), pearl.getY(), pearl.getZ(), RegisterSounds.ITEM_ENDER_PEARL_LAND, owner.getSoundSource(), 0.6F, pitch);
+				FrozenSoundPackets.createLocalPlayerSound(owner, RegisterSounds.ITEM_ENDER_PEARL_LAND, 0.6F, pitch);
+			}
+			if (!owner.isSilent()) {
+				float pitch = 0.9F + (pearl.level().random.nextFloat() * 0.2F);
+				pearl.level().playSound(owner, pearl.getX(), pearl.getY(), pearl.getZ(), SoundEvents.CHORUS_FRUIT_TELEPORT, owner.getSoundSource(), 0.4F, pitch);
+				FrozenSoundPackets.createLocalPlayerSound(owner, SoundEvents.CHORUS_FRUIT_TELEPORT, 0.4F, pitch);
+			}
+		}
+	}
+
+	@Inject(
+		method = "onHit",
+		at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;resetFallDistance()V", ordinal = 1),
+		locals = LocalCapture.CAPTURE_FAILHARD
+	)
+	public void wilderWild$onHitWithoutServerPlayer(HitResult result, CallbackInfo info, Entity owner) {
 		if (ItemConfig.get().projectileLandingSounds.enderPearlLandingSounds) {
 			ThrownEnderpearl pearl = ThrownEnderpearl.class.cast(this);
-			if (!pearl.level().isClientSide) {
-				Entity owner = pearl.getOwner();
-
-				if (owner instanceof ServerPlayer player) {
-					if (!pearl.isSilent()) {
-						float pitch = 0.9F + (pearl.level().random.nextFloat() * 0.2F);
-						pearl.level().playSound(player, pearl.getX(), pearl.getY(), pearl.getZ(), RegisterSounds.ITEM_ENDER_PEARL_LAND, SoundSource.NEUTRAL, 0.6F, pitch);
-						FrozenSoundPackets.createLocalPlayerSound(player, RegisterSounds.ITEM_ENDER_PEARL_LAND, 0.6F, pitch);
-					}
-					if (!player.isSilent()) {
-						float pitch = 0.9F + (pearl.level().random.nextFloat() * 0.2F);
-						pearl.level().playSound(player, pearl.getX(), pearl.getY(), pearl.getZ(), SoundEvents.CHORUS_FRUIT_TELEPORT, SoundSource.NEUTRAL, 0.4F, pitch);
-						FrozenSoundPackets.createLocalPlayerSound(player, SoundEvents.CHORUS_FRUIT_TELEPORT, 0.4F, pitch);
-					}
-				} else {
-					if (!pearl.isSilent()) {
-						pearl.level().playSound(null, pearl.getX(), pearl.getY(), pearl.getZ(), RegisterSounds.ITEM_ENDER_PEARL_LAND, SoundSource.NEUTRAL, 0.6F, 0.85F + (pearl.level().random.nextFloat() * 0.2F));
-					}
-					if (owner != null && !owner.isSilent()) {
-						pearl.level().playSound(null, pearl.getX(), pearl.getY(), pearl.getZ(), SoundEvents.CHORUS_FRUIT_TELEPORT, SoundSource.NEUTRAL, 0.4F, 0.85F + (pearl.level().random.nextFloat() * 0.2F));
-					}
-				}
+			if (!pearl.isSilent()) {
+				pearl.level().playSound(null, pearl.getX(), pearl.getY(), pearl.getZ(), RegisterSounds.ITEM_ENDER_PEARL_LAND, owner.getSoundSource(), 0.6F, 0.85F + (pearl.level().random.nextFloat() * 0.2F));
+			}
+			if (owner != null && !owner.isSilent()) {
+				pearl.level().playSound(null, pearl.getX(), pearl.getY(), pearl.getZ(), SoundEvents.CHORUS_FRUIT_TELEPORT, owner.getSoundSource(), 0.4F, 0.85F + (pearl.level().random.nextFloat() * 0.2F));
 			}
 		}
 	}